tumor-segmentation/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
segmentation-models-pytorch
pytorch-lightning
matplotlib
numpy
scikit-learn
wandb
albumentations
opencv-python
Pillow